Pass hooks to Slot Fills as Fill Props to avoid imports from base-context
