+++ assets/js/base/components/cart-checkout/shipping-rates-control/index.js
@@ -6,14 +6,20 @@ import { useEffect } from '@wordpress/element';
 import PropTypes from 'prop-types';
 import { speak } from '@wordpress/a11y';
 import LoadingMask from '@woocommerce/base-components/loading-mask';
-import {
-	ShippingRatesControlPackage,
-	ExperimentalOrderShippingPackages,
-} from '@woocommerce/blocks-checkout';
+import { ExperimentalOrderShippingPackages } from '@woocommerce/blocks-checkout';
 import {
 	getShippingRatesPackageCount,
 	getShippingRatesRateCount,
 } from '@woocommerce/base-utils';
+import {
+	useSelectShippingRate,
+	useStoreCart,
+} from '@woocommerce/base-context/hooks';
+
+/**
+ * Internal dependencies
+ */
+import Packages from './packages';
 
 /**
  * @typedef {import('react')} React
@@ -98,57 +104,19 @@ const ShippingRatesControl = ( {
 				collapsible={ collapsible }
 				noResultsMessage={ noResultsMessage }
 				renderOption={ renderOption }
+				useStoreCart={ useStoreCart }
+				useSelectShippingRate={ useSelectShippingRate }
 			/>
 			<ExperimentalOrderShippingPackages>
 				<Packages
 					packages={ shippingRates }
 					noResultsMessage={ noResultsMessage }
 					renderOption={ renderOption }
+					useSelectShippingRate={ useSelectShippingRate }
 				/>
 			</ExperimentalOrderShippingPackages>
 		</LoadingMask>
 	);
-};
-
-/**
- * Renders multiple packages within the slotfill.
- *
- * @param {Object} props Incoming props.
- * @param {Array} props.packages Array of packages.
- * @param {React.ReactElement} props.noResultsMessage Rendered when there are no rates in a package.
- * @param {boolean} props.collapsible If the package should be rendered as a
- * collapsible panel.
- * @param {boolean} props.collapse If the panel should be collapsed by default,
- * only works if collapsible is true.
- * @param {boolean} props.showItems If we should items below the package name.
- * @param {Function} [props.renderOption] Function to render a shipping rate.
- * @return {React.ReactElement|Array|null} Rendered components.
- */
-const Packages = ( {
-	packages,
-	collapse,
-	showItems,
-	collapsible,
-	noResultsMessage,
-	renderOption,
-} ) => {
-	// If there are no packages, return nothing.
-	if ( ! packages.length ) {
-		return null;
-	}
-
-	return packages.map( ( { package_id: packageId, ...packageData } ) => (
-		<ShippingRatesControlPackage
-			key={ packageId }
-			packageId={ packageId }
-			packageData={ packageData }
-			collapsible={ collapsible }
-			collapse={ collapse }
-			showItems={ showItems }
-			noResultsMessage={ noResultsMessage }
-			renderOption={ renderOption }
-		/>
-	) );
 };
 
 ShippingRatesControl.propTypes = {
