+++ packages/checkout/order-shipping-packages/index.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import { useStoreCart } from '@woocommerce/base-context/hooks';
 
 /**
  * Internal dependencies
@@ -15,7 +14,14 @@ const {
 	Slot: OrderShippingPackagesSlot,
 } = createSlotFill( slotName );
 
-const Slot = ( { className, collapsible, noResultsMessage, renderOption } ) => {
+const Slot = ( {
+	className,
+	collapsible,
+	noResultsMessage,
+	renderOption,
+	useStoreCart,
+	useSelectShippingRate,
+} ) => {
 	// We need to pluck out receiveCart.
 	// eslint-disable-next-line no-unused-vars
 	const { extensions, receiveCart, ...cart } = useStoreCart();
@@ -35,6 +41,7 @@ const Slot = ( { className, collapsible, noResultsMessage, renderOption } ) => {
 				renderOption,
 				extensions,
 				cart,
+				useSelectShippingRate,
 			} }
 		/>
 	);
